# remove support for congruence closure computation