Supporting congruence closure requires the `unioned_keys` iterator,
which in turn requires maintaining the parent/sibling pointers. This
grows the size of nodes, which in turn leads to a 2% slowdown in
various rustc test cases (e.g., the `coercions` benchmark in
perf.rust-lang.org). Given that this feature is not used in chalk or
anywhere else, no reason to support it.

Cargo bench measurements (3 runs each):

before
test unify::tests::big_array_bench_InPlace             ... bench:   1,335,817 ns/iter
test unify::tests::big_array_bench_InPlace             ... bench:   1,030,749 ns/iter
test unify::tests::big_array_bench_InPlace             ... bench:   1,339,391 ns/iter


after
test unify::tests::big_array_bench_InPlace             ... bench:     918,518 ns/iter
test unify::tests::big_array_bench_InPlace             ... bench:   1,190,575 ns/iter
test unify::tests::big_array_bench_InPlace             ... bench:     918,214 ns/iter

A bit uneven but clearly a win.